Add files via upload
In this project, I have analysed data about the number of cases per country each year from 1990 to 2019. The dataset only included a column with the region, which could also include wider-than-country regions likee the Middle East, Sub-Saharan Africa, North America, or the Caribbean. I order to have clean calculations (and not duplicates, like summing cases in Tanzania and South Africa, but then Sub-Saharan Africa, too), I created, using a dictionary of lists only including , an additional column with the continent to which each country belongs. Those rows without a continent were labelled "Not a country", and then filtered out when creating a secondary dataframe df_countries. I used df_countries for my calculations: I calculated the average cases per continent, both separated by years and of all the years; I calculated top 10s worldwide in different years, and top 5s in each continent, and I tracked the rank of Bangladesh throughout the years, both worldwide and among Asian countries only.

For countries in South America, I created another column indicating whether, based on the country's geographical condition, it is a high-risk or a low-risk country (based on humidity and presence of swamps and rainforests), and found that countries labelled as maximum risk were often lower in the rank of cases than high risk ones. For example, Colombia and Brazil, both labelled high-risk, came higher than Venezuela and Guyana, both labelled as maximum risk - although Venezuela, which had position 7 in 2000, rapidly came to the top three after 2005. Doing some research, I related this to changes in governance of Venezuela, as well as Colombia and Brazil having stronger healthcare system more concerned with malaria - which also do result in more reporting and diagnosing of the illness. Thus, my analysis allowed me to connect the data to political situations going on in those countries.

For Bangladesh, in tracking its ranking position, I found a massive improvement in its protection against malaria, going from position 5 in 1990 to position 52 in 2019, and its malaria cases going from 36419 to 110 in the same time period. Similarly to South America, I was able to link this to political events in the country and healthcare programmes against malaria, as well as collaborations with India and Myanmar. 

[The complete notes are in the Python code as comments]

Finally, I created a Power BI dashboard for which I used CSV files produced with Python, mainly df_countries. For this, I took the WHO's categorisation for level of risk of malaria in a ocuntry: 0 cases = malaria-free; 1<=X<10 cases = low-risk; 10<=X<100 = moderate risk; X>=100 = high-risk. I used this to categorise in new columns the status of each country in 1990, 2005, and 2019, to then create graphs showing me the counts per continent. Additionally, I created a donnut chart showing showing the portion of the world in each category in 2019, and have a bar chart showing the number of cases of malaria per country in 1990, 2000, and 2010.